ver2

+++ lab6_1.cpp
@@ -2,10 +2,10 @@
 using namespace std;
 
 int main() {
-    float a  =  20.0, b = 30.0;
-    float c = 4, output;
+    int a  =  20, b = 30;
+    double c = 4, output;
 
-    output = a - b/c;
+    output = (double) a - b/c;
     cout << "output: " << output << endl;
     return 0;
 }
